清理 clang-tidy 告警并改进若干错误处理与类型转换
- read_script: 用带错误检查的 fseek 替代 rewind，增强鲁棒性。
- resolve: 拆分 IPv6 / IPv4 分支，保持逻辑等价的同时提高可读性。
- tucrypto: 去除无意义的 uint32_t 强转。
- tuctl: 移除未使用头文件，统一 if_nametoindex 的类型转换，调整 auto_pin 接口命名并删除重复声明。
- ratelimiter: 统一 ULL 常量后缀，简化 key 比较与令牌桶逻辑写法。
- tutu_csum_fixup: 为 IP/ICMP 长度增加显式类型转换，复用已计算长度，去掉冗余 cast，修复多处 clang-tidy 报警。

+++ tuctl/tuctl.c
@@ -1,7 +1,6 @@
 #include <bpf/bpf.h>
 #include <bpf/libbpf.h>
 
-#include <ctype.h>
 #include <getopt.h>
 #include <linux/if_link.h>
 #include <linux/limits.h>
@@ -1566,8 +1565,7 @@ int cmd_load(int argc, char **argv) {
     int         ifindex;
 
     log_info("Configuring TC for interface: %s", cur_iface);
-
-    ifindex = if_nametoindex(cur_iface);
+    ifindex = (int) if_nametoindex(cur_iface);
     if (ifindex == 0) {
       log_error("Invalid interface: %s", cur_iface);
       continue;
@@ -1666,7 +1664,7 @@ int cmd_unload(int argc, char **argv) {
     int         ifindex;
 
     log_info("Deconfiguring TC for interface: %s", cur_iface);
-    ifindex = if_nametoindex(cur_iface);
+    ifindex = (int) if_nametoindex(cur_iface);
     if (ifindex == 0) {
       log_error("Invalid interface: %s", cur_iface);
       err = -ENOENT;
